Code complete next step is adding two or more junit test on the viewmodel

+++ app/src/main/java/com/example/techradar/model/Content.kt
@@ -10,7 +10,7 @@ data class Content(
     val phone: String,
     val email: String,
     val birthday: String,
-    val wage: Int,
+    val wage: Double,
     val note: String,
     var favorite: Boolean,
     var picture: String?
