+++ app/src/main/java/com/example/techradar/room/dao/ListDao.kt
@@ -33,7 +33,6 @@ interface ListDao {
     fun getAllFavoriteUsers(): Flow<List<ListDto>>
 
 
-   // @Query("UPDATE list SET list_name = COALESCE(:listName, list_name), list_firstname = COALESCE(:listFirstname, list_firstname), list_phone = COALESCE(:listPhone, list_phone), list_email = COALESCE(:listEmail, list_email), list_birthday = COALESCE(:listBirthday, list_birthday), list_wage = COALESCE(:listWage, list_wage), list_note = COALESCE(:listNote, list_note), list_favorite = COALESCE(:listFavorite, list_favorite), list_picture = COALESCE(:listPicture, list_picture) WHERE user_Id = :id")
 
     @Query("""
         UPDATE list
@@ -56,7 +55,7 @@ interface ListDao {
         listPhone: String?,
         listEmail: String?,
         listBirthday: String?,
-        listWage: Int?,
+        listWage: Double?,
         listNote: String?,
         listFavorite: Boolean?,
         listPicture: String?
